Move database utility functions into separate modules
